requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain_community
langchain-text-splitters
langchain_google_genai
langchain-chroma
chromadb
tqdm
python-dotenv
httpx
PyMuPDF
google-generativeai
pydantic
typing_extensions
setuptools
python-multipart
